requirements.txt
streamlit
pymupdf
pytesseract
pillow
sentence-transformers
torch
langchain
langchain-core
langchain-community
langchain-text-splitters
chromadb